add apt.conf.d Binary::apt::APT::Keep-Downloaded-Packages true
... since by default APT doesn't keep the downloaded .deb files in
/var/cache/apt/archives; apt-get does, but apt doesn't. Since
wasta-offline-setup depends on these packages, it makes sense
to configure apt to keep them around - which matches the
historical behavior of apt-get. (At a time when 1TB hard drives
are the norm, saving space seems like the wrong direction to
have gone in the first place - but of course we are much more
bandwidth sensitive than other parts of the world.)

Added some hints to the log file to re-inforce this, since I
was fairly confused for a while when I started transitioning
to apt. apt-get is a better choice for scripting since its
API and output are stable, while APT is designed for
interactive use and not programmatic use.