non-distributables: add folder hint to template
Installing the local fonts was failing because it
was expecting EVERY file in that folder to be an
installable font. This creating an error status
from some of the other non-font files, and thus,
seeing an error, it fell back to using the
Internet for everything.

Add a folder-name to the template path as a hint
to the end user to create that folder and
organize their downloaded files nicely.