2.1.14: support Mint 21.x vanessa and increase compat to 12
Compat 9 is giving lots of deprication warnings in 20.04 where
we are all building. Ubuntu documentation indicates 12 is the
recommended level for 20.04, but "man debhelper" says 13 is now
recommended (and 11 is highly discouraged).  18.04 recommended 11
and knew about 12 as experiental.

So changing the compat level to 12 should in theory work for
still-in-support 18.04, is not cutting edge, and has no black
marks against it. So this sounds like the most reasonable level
to bump up to.

P.S. 22.04 recommends 13 and knows about 14 and 15.