2.4.0: Don't share repo folders between Mint and Ubuntu
This will still work with older wasta-offline clients,
because they checked for BOTH choices.
This new version is required for the new client.

If wasta-offline is used at all for Linux Mint,
the likelihood of someone sharing the offline cache between
Ubuntu and Mint machines is very high,
since Wasta is based on Ubuntu.

Well, that is just a recipe for disaster.
In my brief use of this with Mint 22.1 xia,
I saw the lsb_release info change to Ubuntu settings,
and that can't be a good thing.

So, just bite the USB-disk-space-usage bullet
and keep entirely separate folders for Mint and Ubuntu.